category/games/2026/02/20260209-022101-vector-circus-charlie-jump/pyproject.toml
[project]
name = "vector-circus-charlie-jump"
version = "0.1.0"
description = "Circus-themed arcade jumping game"
requires-python = ">=3.12"
dependencies = [
    "pygame-ce>=2.5.0",
]

[project.scripts]
start = "main:Game"
